ai_telegram_bot/requirements.txt
python-telegram-bot==21.6
requests
openai
aiohttp
psycopg2-binary